# cmd/api-server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.Candidate:
    properties:
      batch_id:
        type: string
      description:
        type: string
      discovered_at:
        type: string
      id:
        type: string
      ingestion_method:
        type: string
      published_at:
        type: string
      source_abbr:
        type: string
      title:
        type: string
      trace_id:
        type: string
      url:
        type: string
    type: object
  api.Content:
    properties:
      author:
        type: string
      batch_id:
        type: string
      candidate_id:
        type: string
      content:
        type: string
      fetched_at:
        type: string
      id:
        type: string
      published_at:
        type: string
      source_abbr:
        type: string
      title:
        type: string
      trace_id:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  api.FetchProgressResponse:
    properties:
      already_complete:
        $ref: '#/definitions/api.FetchProgressStatus'
      completed:
        $ref: '#/definitions/api.FetchProgressStatus'
      failed:
        $ref: '#/definitions/api.FetchProgressStatus'
      fetch_id:
        type: string
      pending:
        $ref: '#/definitions/api.FetchProgressStatus'
      running:
        $ref: '#/definitions/api.FetchProgressStatus'
      terminal:
        type: boolean
      total:
        type: integer
    type: object
  api.FetchProgressStatus:
    properties:
      candidate_ids:
        items:
          type: string
        type: array
      count:
        type: integer
    type: object
  api.ListCandidatesResponse:
    properties:
      count:
        type: integer
      items:
        items:
          $ref: '#/definitions/api.Candidate'
        type: array
      limit:
        type: integer
      offset:
        type: integer
    type: object
...
  api.PageFetchItem:
    properties:
      candidate_id:
        type: string
      status:
        type: string
    type: object
  api.PageFetchRequest:
    properties:
      candidate_ids:
        items:
          type: string
        type: array
    type: object
  api.PageFetchResponse:
    properties:
      fetch_id:
        type: string
      items:
        items:
          $ref: '#/definitions/api.PageFetchItem'
        type: array
    type: object
...
  api.PostStatusPayload:
    properties:
      level:
        $ref: '#/definitions/obs.HealthLevel'
      message:
        type: string
      service:
        type: string
      timestamp:
        type: string
      uptime:
        type: string
    type: object
  api.StatusRecordResponse:
    properties:
      status:
        type: string
    type: object
...
  obs.HealthLevel:
    enum:
    - STARTING
    - OK
    - WARN
    - ERROR
    type: string
    x-enum-varnames:
    - LevelStarting
    - LevelOK
    - LevelWarn
    - LevelError
  obs.HealthStatus:
    properties:
      level:
        $ref: '#/definitions/obs.HealthLevel'
      message:
        type: string
      timestamp:
        type: string
      uptime:
        type: string
    type: object
...
info:
  contact: {}
  description: User-facing read API for Prism candidate/content data.
  title: Prism API
  version: "0.1"
paths:
...
  /candidates:
    get:
      parameters:
      - description: Keyword matched against title/description (ILIKE)
        in: query
        name: q
        type: string
      - description: Filter by source abbreviation (e.g. dpp, tpp, yahoo)
        in: query
        name: source_abbr
        type: string
      - description: Lower bound on published_at/discovered_at (RFC3339)
        in: query
        name: since
        type: string
      - description: Upper bound on published_at/discovered_at (RFC3339)
        in: query
        name: until
        type: string
      - description: Page size (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Pagination offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ListCandidatesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List candidate article briefs
      tags:
      - candidates
  /contents/{candidate_id}:
    get:
      parameters:
      - description: Candidate UUID
        in: path
        name: candidate_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Content'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get fetched content for a candidate
      tags:
      - contents
  /fetches/{id}:
    get:
      parameters:
      - description: User fetch ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.FetchProgressResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get progress for a user fetch
      tags:
      - fetches
...
  /page_fetch:
    post:
      consumes:
      - application/json
      parameters:
      - description: Candidate IDs to promote to contents
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.PageFetchRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/api.PageFetchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Request full-article fetch for candidates
      tags:
      - candidates
...
  /status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/obs.HealthStatus'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List monitored service statuses
      tags:
      - status
    post:
      consumes:
      - application/json
      parameters:
      - description: Service status payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.PostStatusPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusRecordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Record monitored service status
      tags:
      - status
swagger: "2.0"